requirements.txt
dj-database-url==0.3.0
Django==1.9.1
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
djangorestframework==3.3.2